added Javadoc comments
