+++ src/main/java/frc/aster/RobotContainer.java
@@ -8,22 +8,39 @@ import java.util.function.Supplier;
 
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.wpilibj2.command.DeferredCommand;
 import frc.aster.commands.drive.TurnCommand;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 import frc.aster.subsystems.DriveSubsystem;
+import frc.common.PoseEstimationSubsystem;
 import frc.aster.Constants.ControllerConstants;
-import frc.aster.commands.drive.DefaultDriveCommand;
 import frc.aster.commands.drive.DriveDistanceCommand;
-import frc.robot.subsystems.AprilTagSubsystem;
-import frc.robot.subsystems.PoseEstimationSubsystem;
 import hlib.drive.Pose;
+import hlib.drive.PoseCalculatorWestCoastSimple;
 import hlib.drive.Position;
 
-public class RobotContainer implements frc.robot.util.RobotContainer {
+public class RobotContainer implements frc.common.RobotContainer {
 	DriveSubsystem m_driveSubsystem = new DriveSubsystem();
-	// AprilTagSubsystem m_aprilTagSubsystem = new AprilTagSubsystem();
-	PoseEstimationSubsystem m_poseEstimationSubsystem = new PoseEstimationSubsystem();
+	PoseEstimationSubsystem m_poseEstimationSubsystem = new PoseEstimationSubsystem() {
+		{
+			m_poseEstimator.add(new PoseCalculatorWestCoastSimple(Constants.DriveConstants.kTrackwidthMeters) {
+
+				@Override
+				public State state() {
+					var state = new State(DriveSubsystem.get().getLeftEncoderPosition(),
+							DriveSubsystem.get().getRightEncoderPosition(),
+							Math.toRadians(DriveSubsystem.get().getHeading()));
+					visionTable.getEntry("odometry").setString(state.toString());
+					System.out.println(state);
+					return state;
+				}
+			});
+		}
+
+		@Override
+		public void periodic() {
+			m_poseEstimator.periodic();
+		}
+	};
 	Joystick m_operatorController = new Joystick(ControllerConstants.kOperatorControllerPort);
 
 	public RobotContainer() {
