+++ src/main/java/frc/common/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Provides interfaces and classes applicable across various robots.
+ * 
+ * @author Andrew Hwang (u.andrew.h@gmail.com)
+ * @author Jeong-Hyon Hwang (jhhbrown@gmail.com)
+ */
+package frc.common;
