--- src/main/java/frc/robot/util/RobotContainer.java
@@ -1,8 +1,0 @@
-package frc.robot.util;
-
-import edu.wpi.first.wpilibj2.command.Command;
-
-public interface RobotContainer {
-
-	public Command getAutonomousCommand();
-}
